requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-user-activity-log2==0.0.24
drf-yasg==1.20.0
